* use 32-bit indices for pak files

+++ src/core/asset.c
@@ -25,7 +25,7 @@ void asset_load_pak(byte *asset_pak, const char *file_name)
 #endif
 }
 
-struct AssetHandle asset_make_handle(uint16_t offset, byte *pak_file)
+struct AssetHandle asset_make_handle(uint32_t offset, byte *pak_file)
 {
     struct AssetHandle handle;
     handle.offset = offset;
